#311: Ensure client-rendered errors are recorded
